Move non-entry-points into subfolders
