+++ src/gatsby-shell/comps/SheetTemplateView.tsx
@@ -1,18 +1,18 @@
 import React from 'react'
 import Helmet from 'react-helmet'
-import useSiteContent from '../gatsby-hooks/useSiteContent'
-import { useSheetContext } from './SheetTemplate'
+import useSiteContent from '../../gatsby-hooks/useSiteContent'
+import { useSheetContext } from '../SheetTemplate'
 
 // Components
-import CommentsArea from '../web-comments/CommentsArea'
-import PostContent from '../web-post-content'
-import IntroContent from '../web/components/IntroContent'
-import MainHeading from '../web/components/MainHeading'
-import MiniMarkdown from '../web/components/MiniMarkdown'
-import PreFooter from '../web/components/PreFooter'
-import RelatedPostsArea from '../web/components/RelatedPostsArea'
-import SearchFooter from '../web/components/SearchFooter'
-import TopNav from '../web/components/TopNav'
+import CommentsArea from '../../web-comments/CommentsArea'
+import PostContent from '../../web-post-content/PostContent'
+import IntroContent from '../../web/components/IntroContent'
+import MainHeading from '../../web/components/MainHeading'
+import MiniMarkdown from '../../web/components/MiniMarkdown'
+import PreFooter from '../../web/components/PreFooter'
+import RelatedPostsArea from '../../web/components/RelatedPostsArea'
+import SearchFooter from '../../web/components/SearchFooter'
+import TopNav from '../../web/components/TopNav'
 import CommonHead from './CommonHead'
 
 /**
