+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { graphql, useStaticQuery } from 'gatsby'
 import React from 'react'
-import Layout from '../gatsby-shell/Layout'
+import Layout from '../gatsby-shell/comps/Layout'
 import { groupByCategory, toSiteLinks } from '../helpers/site_page'
 import { AllSitePage } from '../types/types'
 import RootPage from '../web/components/RootPage'
