--- src/styleguides/home.mdx
@@ -1,7 +1,0 @@
-# Home
-
-Hey these are same styleguides
-
-## Yes
-
-That's true
