--- src/web-icons/README.md
@@ -1,1 +1,0 @@
-Icon SVG's taken from clarity icons. See: https://vmware.github.io/clarity/icons/icon-sets
