--- src/web-search/README.md
@@ -1,1 +1,0 @@
-Client implementation for elasticlunr.
