+++ src/web-search/comps/SearchItem.tsx
@@ -2,7 +2,7 @@ import { Link } from 'gatsby'
 import React from 'react'
 import CSS from './SearchItem.module.css'
 
-import { SearchPageItem } from '../types/types'
+import { SearchPageItem } from '../../types/types'
 
 export interface Props {
   page: SearchPageItem
